LPS-108386 regen

+++ modules/apps/account/account-service/src/main/java/com/liferay/account/service/base/AccountEntryLocalServiceBaseImpl.java
@@ -279,7 +279,8 @@ public abstract class AccountEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return accountEntryPersistence.create(primaryKeyObj);
+		return accountEntryPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
