+++ modules/apps/account/account-service/src/main/java/com/liferay/account/service/base/AccountEntryUserRelLocalServiceBaseImpl.java
@@ -297,7 +297,8 @@ public abstract class AccountEntryUserRelLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return accountEntryUserRelPersistence.create(primaryKeyObj);
+		return accountEntryUserRelPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
