+++ modules/apps/adaptive-media/adaptive-media-image-service/src/main/java/com/liferay/adaptive/media/image/service/base/AMImageEntryLocalServiceBaseImpl.java
@@ -290,7 +290,8 @@ public abstract class AMImageEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return amImageEntryPersistence.create(primaryKeyObj);
+		return amImageEntryPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
