+++ modules/apps/app-builder/app-builder-service/src/main/java/com/liferay/app/builder/service/base/AppBuilderAppDeploymentLocalServiceBaseImpl.java
@@ -300,7 +300,8 @@ public abstract class AppBuilderAppDeploymentLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return appBuilderAppDeploymentPersistence.create(primaryKeyObj);
+		return appBuilderAppDeploymentPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
