+++ modules/apps/archived/chat-service/src/main/java/com/liferay/chat/service/base/EntryLocalServiceBaseImpl.java
@@ -271,7 +271,7 @@ public abstract class EntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return entryPersistence.create(primaryKeyObj);
+		return entryPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
