+++ modules/apps/archived/chat-service/src/main/java/com/liferay/chat/service/base/StatusLocalServiceBaseImpl.java
@@ -272,7 +272,7 @@ public abstract class StatusLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return statusPersistence.create(primaryKeyObj);
+		return statusPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
