+++ modules/apps/archived/mail-reader-service/src/main/java/com/liferay/mail/reader/service/base/AccountLocalServiceBaseImpl.java
@@ -275,7 +275,7 @@ public abstract class AccountLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return accountPersistence.create(primaryKeyObj);
+		return accountPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
