+++ modules/apps/archived/mail-reader-service/src/main/java/com/liferay/mail/reader/service/base/AttachmentLocalServiceBaseImpl.java
@@ -277,7 +277,7 @@ public abstract class AttachmentLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return attachmentPersistence.create(primaryKeyObj);
+		return attachmentPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
