+++ modules/apps/archived/mail-reader-service/src/main/java/com/liferay/mail/reader/service/base/FolderLocalServiceBaseImpl.java
@@ -273,7 +273,7 @@ public abstract class FolderLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return folderPersistence.create(primaryKeyObj);
+		return folderPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
