+++ modules/apps/archived/opensocial-portlet/docroot/WEB-INF/src/com/liferay/opensocial/service/base/GadgetLocalServiceBaseImpl.java
@@ -357,7 +357,7 @@ public abstract class GadgetLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return gadgetPersistence.create(primaryKeyObj);
+		return gadgetPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
