+++ modules/apps/archived/opensocial-portlet/docroot/WEB-INF/src/com/liferay/opensocial/service/base/OAuthTokenLocalServiceBaseImpl.java
@@ -276,7 +276,7 @@ public abstract class OAuthTokenLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return oAuthTokenPersistence.create(primaryKeyObj);
+		return oAuthTokenPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
