+++ modules/apps/archived/powwow-portlet/docroot/WEB-INF/src/com/liferay/powwow/service/base/PowwowMeetingLocalServiceBaseImpl.java
@@ -282,7 +282,8 @@ public abstract class PowwowMeetingLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return powwowMeetingPersistence.create(primaryKeyObj);
+		return powwowMeetingPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
