+++ modules/apps/archived/powwow-portlet/docroot/WEB-INF/src/com/liferay/powwow/service/base/PowwowParticipantLocalServiceBaseImpl.java
@@ -287,7 +287,8 @@ public abstract class PowwowParticipantLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return powwowParticipantPersistence.create(primaryKeyObj);
+		return powwowParticipantPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
