+++ modules/apps/archived/tasks-portlet/docroot/WEB-INF/src/com/liferay/tasks/service/base/TasksEntryLocalServiceBaseImpl.java
@@ -281,7 +281,7 @@ public abstract class TasksEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return tasksEntryPersistence.create(primaryKeyObj);
+		return tasksEntryPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
