+++ modules/apps/asset/asset-category-property-service/src/main/java/com/liferay/asset/category/property/service/base/AssetCategoryPropertyLocalServiceBaseImpl.java
@@ -298,7 +298,8 @@ public abstract class AssetCategoryPropertyLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return assetCategoryPropertyPersistence.create(primaryKeyObj);
+		return assetCategoryPropertyPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
