+++ modules/apps/asset/asset-display-page-service/src/main/java/com/liferay/asset/display/page/service/base/AssetDisplayPageEntryLocalServiceBaseImpl.java
@@ -415,7 +415,8 @@ public abstract class AssetDisplayPageEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return assetDisplayPageEntryPersistence.create(primaryKeyObj);
+		return assetDisplayPageEntryPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
