+++ modules/apps/asset/asset-list-service/src/main/java/com/liferay/asset/list/service/base/AssetListEntryLocalServiceBaseImpl.java
@@ -370,7 +370,8 @@ public abstract class AssetListEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return assetListEntryPersistence.create(primaryKeyObj);
+		return assetListEntryPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
