+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/service/base/BatchEngineImportTaskLocalServiceBaseImpl.java
@@ -396,7 +396,8 @@ public abstract class BatchEngineImportTaskLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return batchEngineImportTaskPersistence.create(primaryKeyObj);
+		return batchEngineImportTaskPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
