+++ modules/apps/blogs/blogs-service/src/main/java/com/liferay/blogs/service/base/BlogsStatsUserLocalServiceBaseImpl.java
@@ -279,7 +279,8 @@ public abstract class BlogsStatsUserLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return blogsStatsUserPersistence.create(primaryKeyObj);
+		return blogsStatsUserPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
