+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/service/base/CTCollectionLocalServiceBaseImpl.java
@@ -281,7 +281,8 @@ public abstract class CTCollectionLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return ctCollectionPersistence.create(primaryKeyObj);
+		return ctCollectionPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
