+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/service/base/CTEntryLocalServiceBaseImpl.java
@@ -273,7 +273,7 @@ public abstract class CTEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return ctEntryPersistence.create(primaryKeyObj);
+		return ctEntryPersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
