+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/service/base/CTMessageLocalServiceBaseImpl.java
@@ -272,7 +272,7 @@ public abstract class CTMessageLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return ctMessagePersistence.create(primaryKeyObj);
+		return ctMessagePersistence.create(((Long)primaryKeyObj).longValue());
 	}
 
 	/**
