+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/service/base/CTPreferencesLocalServiceBaseImpl.java
@@ -276,7 +276,8 @@ public abstract class CTPreferencesLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return ctPreferencesPersistence.create(primaryKeyObj);
+		return ctPreferencesPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
