+++ modules/apps/changeset/changeset-service/src/main/java/com/liferay/changeset/service/base/ChangesetEntryLocalServiceBaseImpl.java
@@ -277,7 +277,8 @@ public abstract class ChangesetEntryLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return changesetEntryPersistence.create(primaryKeyObj);
+		return changesetEntryPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
