+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/service/base/DEDataDefinitionFieldLinkLocalServiceBaseImpl.java
@@ -316,7 +316,8 @@ public abstract class DEDataDefinitionFieldLinkLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return deDataDefinitionFieldLinkPersistence.create(primaryKeyObj);
+		return deDataDefinitionFieldLinkPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
