+++ modules/apps/data-engine/data-engine-service/src/main/java/com/liferay/data/engine/service/base/DEDataListViewLocalServiceBaseImpl.java
@@ -362,7 +362,8 @@ public abstract class DEDataListViewLocalServiceBaseImpl
 	public PersistedModel createPersistedModel(Serializable primaryKeyObj)
 		throws PortalException {
 
-		return deDataListViewPersistence.create(primaryKeyObj);
+		return deDataListViewPersistence.create(
+			((Long)primaryKeyObj).longValue());
 	}
 
 	/**
